finish group and subject features

+++ src/subject/subject.service.ts
@@ -6,9 +6,9 @@ import { Subject } from 'generated/prisma';
 
 @Injectable()
 export class SubjectService {
-  constructor( private prisma: PrismaService,
-  ) {}
-   async  create(createSubjectDto: CreateSubjectDto , userId : number) :Promise<Subject>{
+  constructor(private prisma: PrismaService,
+  ) { }
+  async create(createSubjectDto: CreateSubjectDto, userId: number): Promise<Subject> {
     let user = await this.prisma.user.findUnique({
       where: {
         id: userId,
@@ -17,7 +17,7 @@ export class SubjectService {
 
     let verify = await this.prisma.subject.findFirst({
       where: {
-        title: createSubjectDto.title,  
+        title: createSubjectDto.title,
       },
     });
 
@@ -29,27 +29,74 @@ export class SubjectService {
       data: {
         title: createSubjectDto.title,
         description: createSubjectDto.description,
-        teacherId : userId,
+        teacherId: userId,
       },
     });
     return subject;
-    
-    
+
+
   }
 
-  
+
 
   async findOne(id: number) {
+    let subject = await this.prisma.subject.findFirst({
+      where: {
+        id 
+      },
+      include: {
+        
+        teacher: {
+          select: {
+            password: false,
+            id: true,
+            email: true,
+            name: true,
+            
+          },
+        },
+      },
+    });
+    if (!subject) {
+      throw new BadRequestException('Subject not found');
+    }
+    return subject;
+  }
+
+  async findall(){
+    let subjects = await this.prisma.subject.findMany({
+      include: {
+        teacher: {
+          select: {
+            password: false,
+            id: true,
+            email: true,
+            name: true,
+            
+          },
+        },
+      },
+      
+    });
+    return subjects;
+  }
+
+ 
+
+  async update(id: number, updateSubjectDto: UpdateSubjectDto , userId: number): Promise<Subject> {
+
     let subject = await this.prisma.subject.findUnique({
       where: {
         id: id,
       },
     });
-    return subject;
-  }
-
-  async update(id: number, updateSubjectDto: UpdateSubjectDto) : Promise<Subject> {
-    let subject = await this.prisma.subject.update({
+    if (!subject) {
+      throw new BadRequestException('Subject not found');
+    }
+    if (subject.teacherId !== userId) {
+      throw new BadRequestException('You are not authorized to update this subject');
+    }
+    let updatedSubject = await this.prisma.subject.update({
       where: {
         id: id,
       },
@@ -58,17 +105,27 @@ export class SubjectService {
         description: updateSubjectDto.description,
       },
     });
-    return subject;
-    
+    return updatedSubject;
+
   }
 
-  async remove(id: number) {
-    let subject = await this.prisma.subject.delete({
+  async remove(id: number , userId: number) {
+    let subject = await this.prisma.subject.findUnique({
       where: {
         id: id,
       },
     });
-    return subject;
-    
+    if (!subject) {
+      throw new BadRequestException('Subject not found');
+    }
+    if (subject.teacherId !== userId) {
+      throw new BadRequestException('You are not authorized to delete this subject');
+    }
+    let deletedSubject = await this.prisma.subject.delete({
+      where: {
+        id: id,
+      },
+    });
+    return { message: 'Subject deleted successfully'};
   }
 }
